+++ IBM_android_assignment/app/src/main/java/com/ibm/rides/data/model/domain/Vehicle.kt
@@ -1,4 +1,4 @@
-package com.ibm.rides.model
+package com.ibm.rides.data.model.domain
 import com.google.gson.annotations.SerializedName
 
 data class Vehicle(
