+++ IBM_android_assignment/app/src/main/java/com/ibm/rides/data/network/ApiClient.kt
@@ -1,4 +1,4 @@
-package com.ibm.rides.api
+package com.ibm.rides.data.network
 
 import com.ibm.rides.BuildConfig
 import retrofit2.Retrofit
